Fix type conversion error for attr_count (uint to int cast)

+++ src/RazerController.Native/RazerDevice.cs
@@ -60,7 +60,7 @@ public class RazerDevice
         
         // The native attr_list array has a maximum size of 64 elements
         const int MAX_ATTR_LIST_SIZE = 64;
-        int actualCount = Math.Min(_device.attr_count, MAX_ATTR_LIST_SIZE);
+        int actualCount = Math.Min((int)_device.attr_count, MAX_ATTR_LIST_SIZE);
         Logger.Debug($"Loading attributes: attr_count={_device.attr_count}, attr_list pointer={_device.attr_list:X}, reading {actualCount} attributes");
         
         for (int i = 0; i < actualCount; i++)
